+++ src/main/java/atdd/path/application/UserService.java
@@ -1,7 +1,10 @@
 package atdd.path.application;
 
+import atdd.path.application.dto.LoginRequestView;
+import atdd.path.application.dto.LoginResponseView;
 import atdd.path.application.dto.UserRequestView;
 import atdd.path.application.dto.UserResponseView;
+import atdd.path.application.exception.FailedLoginException;
 import atdd.path.application.exception.InvalidJwtAuthenticationException;
 import atdd.path.domain.User;
 import atdd.path.repository.UserRepository;
@@ -9,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.Objects;
 
 @Service
 public class UserService {
@@ -30,7 +34,7 @@ public class UserService {
         userRepository.deleteById(id);
     }
 
-    public UserResponseView retrieveUser(HttpServletRequest request)  {
+    public UserResponseView retrieveUser(HttpServletRequest request) {
         String email = this.extractEmail(request);
 
         User persistUser = userRepository.findUserByEmail(email);
@@ -43,5 +47,19 @@ public class UserService {
             throw new InvalidJwtAuthenticationException("invalid token");
         }
         return jwtTokenProvider.getUserEmail(token);
+    }
+
+
+    public LoginResponseView login(LoginRequestView loginRequestView) {
+        User userInfo = userRepository.findUserByEmail(loginRequestView.getEmail());
+
+        if (Objects.isNull(userInfo) || !userInfo.getPassword().equals(loginRequestView.getPassword())) {
+            throw new FailedLoginException();
+        }
+
+        String token = jwtTokenProvider.createToken(loginRequestView.getEmail());
+        return LoginResponseView.builder()
+                .accessToken(token)
+                .build();
     }
 }
