+++ src/main/java/atdd/path/application/exception/FailedLoginException.java
@@ -1,0 +1,4 @@
+package atdd.path.application.exception;
+
+public class FailedLoginException extends RuntimeException {
+}
