+++ src/main/java/atdd/path/web/LoginController.java
@@ -1,6 +1,6 @@
 package atdd.path.web;
 
-import atdd.path.application.LoginService;
+import atdd.path.application.UserService;
 import atdd.path.application.dto.LoginRequestView;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,14 +9,14 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("login")
 public class LoginController {
 
-    private final LoginService loginService;
+    private final UserService userService;
 
-    public LoginController(LoginService loginService) {
-        this.loginService = loginService;
+    public LoginController(UserService userService) {
+        this.userService = userService;
     }
 
     @PostMapping
     public ResponseEntity login(@RequestBody LoginRequestView loginRequestView) {
-        return ResponseEntity.ok(loginService.login(loginRequestView));
+        return ResponseEntity.ok(userService.login(loginRequestView));
     }
 }
